// entrypoints/types/VideoData.ts
export interface VideoDataRootObject {
    meta?: Meta;
    data?: Data;
}

interface Data {
    metadata: Metadata;
    googleTagManager: GoogleTagManager;
    response: Response;
}

interface Response {
    // ads と category は不明
    ads: any;
    category: null;
    channel: Channel | null;
    client: Client;
    comment: Comment;
    // 多分もう復活しないのでnull
    community: null;
    easyComment: EasyComment;
    external: External;
    genre: Genre;
    marquee: Marquee;
    media: Media;
    okReason: string;
    owner: Owner;
    payment: Payment;
    pcWatchPage: PcWatchPage;
    player: Player;
    // 多分ポイントで買うやつ。復活待ち
    ppv: any;
    ranking: Ranking;
    series: Series;
    smartphone: null;
    system: System;
    tag: Tag;
    video: VideoInfo;
    videoAds: VideoAds;
    // ライブ公開のオブジェクト？
    videoLive: any;
    viewer: ViewerInfo;
    waku: Waku;
}
// ...
interface GoogleTagManager {
    niconico: Niconico;
    channel: null;
}

interface Niconico {
    user: User;
    content: Content;
}

interface Content {
    player_type: string;
    genre: string;
    content_type: string;
}

interface User {
    user_id: string;
    login_status: string;
    member_status: string;
    ui_area: string;
    ui_lang: string;
}

interface Metadata {
    title: string;
    linkTags: LinkTag[];
    metaTags: MetaTag[];
    jsonLds: JsonLd[];
}
// ...
interface Meta {
    status: number;
    code: string;
}